feat(tests): p2p_large::large_sync: replace payloads MerkleMap index type with u8
This way, its better spreaded on all levels, when the actual indexes are only in
range of (0..200)

+++ packages/kolme-test/src/p2p_large.rs
@@ -13,7 +13,7 @@ pub struct SampleKolmeApp {
 #[derive(Clone, Debug)]
 pub struct SampleState {
     next_hi: u64,
-    payloads: MerkleMap<u64, usize>,
+    payloads: MerkleMap<u8, usize>,
 }
 
 impl MerkleSerialize for SampleState {
@@ -100,7 +100,7 @@ impl KolmeApp for SampleKolmeApp {
         match msg {
             SampleMessage::SayHi { payload } => {
                 let state = ctx.state_mut();
-                let idx = state.next_hi;
+                let idx = state.next_hi as u8;
                 state.next_hi += 1;
                 state.payloads.insert(idx, payload.len());
             }
